Correct write detection for conditionals.
This commit changes dependency detection for conditionals so that writes which
appear on all branches are labelled as unconditional writes. This commit also
replaces ad-hoc solutions for set operations (union, intersection, etc) with
calls to built-in functions.

This means loop-carried dependency detection is now handled correctly, except
in the case of functions which rely on non-standard evaluation.