Add loop fusion dependency graph generation.
This commit adds full support for generating the dependency graph used in loop
fusion. A notable component of the new graph generation code is the blockify
function, which agglomerates individual code objects into blocks.

A limitation is that antidepdences in loops are not yet handled per variable,
so any antidependence in a loop blocks fusion. This should be straightforward
to fix by adding helper functions to the dependency collector.

A future commit could add more precise dependence tracking (true, anti, or
output) per variable, and add support for subscripted variables.